Revert "Finish Launcher"
