--- .gitignore
@@ -1,3 +1,0 @@
-
-.DS_Store
-.vscode/
