--- 1/canvas/main.js
@@ -1,10 +1,0 @@
-// © 2021 Gavin Morrow
-// EasyJS Canvas
-
-import LineGraph from "./lineGraph/index.js";
-
-const canvas = {
-	LineGraph,
-};
-
-export default canvas;
