--- 1/cookies/cookieConsent/style.css
@@ -1,35 +1,0 @@
-@import "/EasyJS/1/ui/colorScheme/index.js";
-.easyjs-cookie-consent-elem {
-	--br: var(--border-radius, 1em);
-	all: initial;
-	font: inherit;
-	font-size: large;
-	border-radius: var(--br);
-	padding: calc(var(--br) / 1);
-	text-align: center;
-}
-#easyjs-cookie-consent {
-	position: fixed;
-	left: 0;
-	right: 0;
-	bottom: 0;
-	background-color: var(--easyjs-cardBg);
-	box-shadow: 0 0 var(--br) 0 var(--easyjs-subText);
-	color: var(--easyjs-text);
-	border-radius: 0;
-	border-top-left-radius: var(--br);
-	border-top-right-radius: var(--br);
-	opacity: 1;
-	transition: opacity 1s;
-}
-#easyjs-cookie-consent-button {
-	color: var(--easyjs-text);
-	background-color: var(--easyjs-buttonBg);
-	border: none;
-	box-shadow: 0 0 calc(var(--br) / 4) 0 var(--easyjs-subText) inset;
-	transition: box-shadow 1s;
-	margin: var(--br);
-}
-#easyjs-cookie-consent-button:hover, #easyjs-cookie-consent-button:focus {
-	box-shadow: 0 0 calc(var(--br) / 1) 0 var(--easyjs-subText) inset;
-}
