--- 1/cookies/main.js
@@ -1,11 +1,0 @@
-// EasyJS Cookies
-
-import cookieConsent from "./cookieConsent/index.js";
-import Cookie from "./cookie/index.js";
-
-const cookies = {
-	cookieConsent,
-	Cookie,
-};
-
-export default cookies;
