--- 1/math/clamp/index.js
@@ -1,4 +1,0 @@
-// EasyJS Math
-
-const clamp = (min, prefered, max) => Math.max(min, Math.min(prefered, max));
-export default clamp;
