--- 1/math/main.js
@@ -1,11 +1,0 @@
-// EasyJS Math
-
-import random from "./random/index.js";
-import clamp from "./clamp/index.js";
-
-const math = {
-	random,
-	clamp,
-};
-
-export default math;
