--- 1/ui/changeFavicon/index.js
@@ -1,5 +1,0 @@
-// EasyJS UI
-
-const changeFavicon = (link, url) => link.href = url;
-
-export default changeFavicon;
