--- 1/ui/colorScheme/style.css
@@ -1,87 +1,0 @@
-:root {
-	--easyjs-dark-bg:var(--easyjs-dark-bg, #111111);
-	--easyjs-light-bg:var(--easyjs-light-bg, #ffffff);
-	--easyjs-bg:var(--easyjs-bg, #ffffff);
-	--easyjs-inverse-bg:var(--easyjs-inverse-bg, #111111);
-	--easyjs-dark-navBg:var(--easyjs-dark-navBg, #181818);
-	--easyjs-light-navBg:var(--easyjs-light-navBg, #fefffe);
-	--easyjs-navBg:var(--easyjs-navBg, #fefffe);
-	--easyjs-inverse-navBg:var(--easyjs-inverse-navBg, #181818);
-	--easyjs-dark-contrastNavBg:var(--easyjs-dark-contrastNavBg, #858885);
-	--easyjs-light-contrastNavBg:var(--easyjs-light-contrastNavBg, #555855);
-	--easyjs-contrastNavBg:var(--easyjs-contrastNavBg, #555855);
-	--easyjs-inverse-contrastNavBg:var(--easyjs-inverse-contrastNavBg, #858885);
-	--easyjs-dark-cardBg:var(--easyjs-dark-cardBg, #303130);
-	--easyjs-light-cardBg:var(--easyjs-light-cardBg, #efeeef);
-	--easyjs-cardBg:var(--easyjs-cardBg, #efeeef);
-	--easyjs-inverse-cardBg:var(--easyjs-inverse-cardBg, #303130);
-	--easyjs-dark-buttonBg:var(--easyjs-dark-buttonBg, #383938);
-	--easyjs-light-buttonBg:var(--easyjs-light-buttonBg, #e0e1e0);
-	--easyjs-buttonBg:var(--easyjs-buttonBg, #e0e1e0);
-	--easyjs-inverse-buttonBg:var(--easyjs-inverse-buttonBg, #383938);
-	--easyjs-dark-divideBg:var(--easyjs-dark-divideBg, #d3d3d3);
-	--easyjs-light-divideBg:var(--easyjs-light-divideBg, #2c2c2c);
-	--easyjs-divideBg:var(--easyjs-divideBg, #2c2c2c);
-	--easyjs-inverse-divideBg:var(--easyjs-inverse-divideBg, #d3d3d3);
-	--easyjs-dark-alertBg:var(--easyjs-dark-alertBg, #aaaa00);
-	--easyjs-light-alertBg:var(--easyjs-light-alertBg, #ffff00);
-	--easyjs-alertBg:var(--easyjs-alertBg, #ffff00);
-	--easyjs-inverse-alertBg:var(--easyjs-inverse-alertBg, #aaaa00);
-	--easyjs-dark-text:var(--easyjs-dark-text, #ecefed);
-	--easyjs-light-text:var(--easyjs-light-text, #131012);
-	--easyjs-text:var(--easyjs-text, #131012);
-	--easyjs-inverse-text:var(--easyjs-inverse-text, #ecefed);
-	--easyjs-dark-subText:var(--easyjs-dark-subText, #acacac);
-	--easyjs-light-subText:var(--easyjs-light-subText, #535353);
-	--easyjs-subText:var(--easyjs-subText, #535353);
-	--easyjs-inverse-subText:var(--easyjs-inverse-subText, #acacac);
-	--easyjs-dark-link:var(--easyjs-dark-link, #adddad);
-	--easyjs-light-link:var(--easyjs-light-link, #0000ff);
-	--easyjs-link:var(--easyjs-link, #0000ff);
-	--easyjs-inverse-link:var(--easyjs-inverse-link, #adddad);
-}
-
-@media (prefers-color-scheme: dark) {
-	:root {
-		--easyjs-dark-bg:var(--easyjs-dark-bg, #111111);
-		--easyjs-light-bg:var(--easyjs-light-bg, #ffffff);
-		--easyjs-bg:var(--easyjs-bg, #111111);
-		--easyjs-inverse-bg:var(--easyjs-inverse-bg, #ffffff);
-		--easyjs-dark-navBg:var(--easyjs-dark-navBg, #181818);
-		--easyjs-light-navBg:var(--easyjs-light-navBg, #fefffe);
-		--easyjs-navBg:var(--easyjs-navBg, #181818);
-		--easyjs-inverse-navBg:var(--easyjs-inverse-navBg, #fefffe);
-		--easyjs-dark-contrastNavBg:var(--easyjs-dark-contrastNavBg, #858885);
-		--easyjs-light-contrastNavBg:var(--easyjs-light-contrastNavBg, #555855);
-		--easyjs-contrastNavBg:var(--easyjs-contrastNavBg, #858885);
-		--easyjs-inverse-contrastNavBg:var(--easyjs-inverse-contrastNavBg, #555855);
-		--easyjs-dark-cardBg:var(--easyjs-dark-cardBg, #303130);
-		--easyjs-light-cardBg:var(--easyjs-light-cardBg, #efeeef);
-		--easyjs-cardBg:var(--easyjs-cardBg, #303130);
-		--easyjs-inverse-cardBg:var(--easyjs-inverse-cardBg, #efeeef);
-		--easyjs-dark-buttonBg:var(--easyjs-dark-buttonBg, #383938);
-		--easyjs-light-buttonBg:var(--easyjs-light-buttonBg, #e0e1e0);
-		--easyjs-buttonBg:var(--easyjs-buttonBg, #383938);
-		--easyjs-inverse-buttonBg:var(--easyjs-inverse-buttonBg, #e0e1e0);
-		--easyjs-dark-divideBg:var(--easyjs-dark-divideBg, #d3d3d3);
-		--easyjs-light-divideBg:var(--easyjs-light-divideBg, #2c2c2c);
-		--easyjs-divideBg:var(--easyjs-divideBg, #d3d3d3);
-		--easyjs-inverse-divideBg:var(--easyjs-inverse-divideBg, #2c2c2c);
-		--easyjs-dark-alertBg:var(--easyjs-dark-alertBg, #aaaa00);
-		--easyjs-light-alertBg:var(--easyjs-light-alertBg, #ffff00);
-		--easyjs-alertBg:var(--easyjs-alertBg, #aaaa00);
-		--easyjs-inverse-alertBg:var(--easyjs-inverse-alertBg, #ffff00);
-		--easyjs-dark-text:var(--easyjs-dark-text, #ecefed);
-		--easyjs-light-text:var(--easyjs-light-text, #131012);
-		--easyjs-text:var(--easyjs-text, #ecefed);
-		--easyjs-inverse-text:var(--easyjs-inverse-text, #131012);
-		--easyjs-dark-subText:var(--easyjs-dark-subText, #acacac);
-		--easyjs-light-subText:var(--easyjs-light-subText, #535353);
-		--easyjs-subText:var(--easyjs-subText, #acacac);
-		--easyjs-inverse-subText:var(--easyjs-inverse-subText, #535353);
-		--easyjs-dark-link:var(--easyjs-dark-link, #adddad);
-		--easyjs-light-link:var(--easyjs-light-link, #0000ff);
-		--easyjs-link:var(--easyjs-link, #adddad);
-		--easyjs-inverse-link:var(--easyjs-inverse-link, #0000ff);
-	}
-}
