--- 1/ui/main.js
@@ -1,16 +1,0 @@
-// © 2021 Gavin Morrow
-// EasyJS UI
-
-import center from "./center/index.js";
-import changeFavicon from "./changeFavicon/index.js";
-import colorScheme from "./colorScheme/index.js";
-import popup from "./popup/index.js";
-
-const ui = {
-	center,
-	colorScheme,
-	changeFavicon,
-	Popup,
-};
-
-export default ui;
