--- 1/ui/popup/style.css
@@ -1,38 +1,0 @@
-.easyjs-popup-wrapper {
-	--br: var(--border-radius, 0.5em);
-	all: initial;
-	position: fixed;
-	top: 0;
-	left: 0;
-	bottom: 0;
-	right: 0;
-
-	text-align: center;
-	z-index: -1000;
-	transition: opacity 1s;
-	opacity: 0;
-
-	display: flex;
-	justify-content: center;
-	align-items: center;
-
-	font-size: xx-large;
-	font-family: inherit;
-}
-.easyjs-popup {
-	all: inherit;
-	position: relative;
-	text-align: center;
-	color: var(--easyjs-text);
-	background-color: rgba(255, 255, 255, 0.9);
-	box-shadow: 0 0 var(--br) 0 var(--easyjs-text);
-	padding: var(--br);
-	border-radius: var(--br);
-	width: fit-content;
-	height: fit-content;
-}
-@media (prefers-color-scheme: dark) {
-	.easyjs-popup {
-		background-color: rgba(0, 0, 0, 0.9);
-	}
-}
