+++ README.md
@@ -1,3 +1,1 @@
-<h6>v1.0.0</h6>
-<hr>
-Go to the <a href="https://gavinmorrow.github.io/EasyJS/docs/">EasyJS Docs</a> for more information. 
+# easyjs
