--- docs/docs.css
@@ -1,118 +1,0 @@
-.section-title {
-	background-color: var(--easyjs-cardBg);
-	--padding: var(--border-radius);
-	border-radius: var(--border-radius);
-	padding: var(--padding);
-	transition: background-color 0.75;
-}
-.section-title::before {
-	content: "EasyJS ";
-	font-family: var(--font-family-body);
-	font-size: larger;
-}
-.section-title:hover, .section-title:focus {
-	/* Add this when you can collapse sections. */
-	/* background-color: var(--easyjs-buttonBg); */
-}
-
-h2 {
-	padding-bottom: 1vmin;
-	width: fit-content;
-	position: relative;
-}
-h2::after {
-	position: relative;
-	content: "";
-	display: block;
-	width: calc(var(--border-radius) * 4);
-	margin-top: 1vmin;
-	border-bottom: 1px solid var(--easyjs-text);
-	transition: width 1s;
-}
-h2:hover::after, h2:focus::after {
-	width: 100%;
-}
-
-#title {
-	font-family: var(--font-family-body);
-	font-size: xxx-large;
-	margin: 0 0 auto 0;
-	padding: var(--padding);
-	position: fixed;
-	top: 0;
-	width: 100%;
-	transition: box-shadow 0.5s;
-	background-color: var(--easyjs-navBg);
-	z-index: 100;
-}
-#title-inner {
-	display: inline-block;
-	font: smaller var(--font-family-body2);
-}
-
-
-
-#main {
-	--padding-top: var(--border-radius);
-	--padding-right: var(--border-radius);
-	--padding: var(--padding-top) var(--padding-right);
-
-	overflow: scroll;
-	text-align: left;
-	position: relative;
-	z-index: 1;
-
-	padding: 0;
-	margin: 0;
-}
-#main-inner {
-	padding: 0 var(--padding-right);
-	scroll-behavior: smooth;
-}
-
-.example {
-	background-color: var(--easyjs-light-navBg);
-	--padding: var(--border-radius);
-	overflow: scroll;
-	vertical-align: middle;
-	display: inline-block;
-}
-.example.block {
-	display: block;
-	white-space: pre;
-	box-shadow: 0 0 calc(var(--border-radius) / 8) 0 var(--easyjs-light-text) inset;
-	transition: box-shadow 1s;
-	padding: var(--padding);
-	margin: calc(var(--padding) / 2) 0;
-	border-radius: var(--border-radius);
-	color: var(--easyjs-light-text);
-}
-.example.block:hover, .example.block:focus {
-	box-shadow: 0 0 calc(var(--border-radius) / 2) 0 var(--easyjs-light-text) inset;
-}
-.example:not(.block) {
-	background-color: var(--easyjs-light-cardBg);
-	--padding: calc(var(--border-radius) / 2);
-	padding: var(--padding);
-	border-radius: calc(var(--border-radius) / 2);
-}
-
-.linkable-link {
-	display: inline-block;
-	transition: opacity 1s, text-decoration-color 1s;
-	margin-left: var(--border-radius);
-	color: var(--easyjs-link);
-	text-decoration-color: transparent;
-}
-.linkable-link:hover, .linkable-link:focus {
-	text-decoration-color: var(--easyjs-link);
-}
-
-:is(.linkable):target {
-	animation: highlight 6s linear 0s 1 normal none;
-}
-@keyframes highlight {
-	from {}
-	50% { background-color: var(--easyjs-alertBg) }
-	to {}
-}
