--- docs/index.html
@@ -1,74 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-	<meta charset="UTF-8">
-	<meta http-equiv="X-UA-Compatible" content="IE=edge">
-	<meta name="viewport" content="width=device-width, initial-scale=1.0">
-	<title>Docs | EasyJS</title>
-	<link rel="stylesheet" href="style.css">
-	<link rel="stylesheet" href="nav.css">
-	<link rel="stylesheet" href="docs.css">
-	<link rel="stylesheet" href="highlighting.css">
-</head>
-<body>
-	<nav id="nav">
-		<button id="close" class="nav-a">&lt</button>
-		<a href="#overview-title" class="nav-a">Overview</a>
-		<a href="#canvas-title" class="nav-a">Canvas</a>
-		<a href="#math-title" class="nav-a">Math</a>
-		<a href="#string-title" class="nav-a">String</a>
-		<a href="#ui-title" class="nav-a">UI</a>
-	</nav>
-	<article id="main">
-		<h1 id="title">EasyJS <div id="title-inner"></div> <span>v1.0.0</span></h1>
-		<div id="main-inner">
-			<section id="overview" class="section">
-				<h1 id="overview-title" class="section-title linkable">Overview</h1>
-				<h2 id="what-is-easyjs" class="linkable">What is EasyJS?</h2>
-				<div>EasyJS is a JavaScript library for simplifying your code. </div>
-				<h2 id="try-easyjs" class="linkable">Try EasyJS</h2>
-				To try EasyJS, just add it to your page with just a few lines of code:
-				<code class="example block"><span class="tag">&ltscript</span> <span class="attribute">type</span>=<span class="value">"module"</span><span class="tag">></span>
-<span class="keyword">import</span> <span class="variable">EasyJSVersion</span> <span class="keyword">from</span> <span class="string">"http://gavinmorrow.com/easyjs/"</span>;
-
-<span class="comment">// You can set the version number (below) to any valid major version. </span>
-<span class="declaration">const</span> <span class="constant">EasyJS</span> = <span class="declaration">new</span> <span class="function">EasyJSVersion</span>(<span class="number">1</span>); <span class="comment">// This is the actual EasyJS interface. </span>
-
-<span class="comment">// Your code here!</span>
-<span class="tag">&lt/script></span></code>
-				You can also import individual sections or functions/APIs, but there will be more on that later. 
-			</section>
-			<section id="canvas" class="section" data-text-transform="capitalize">
-				<h1 id="canvas-title" class="section-title linkable">Canvas</h1>
-				<h2 id="link-canvas" class="linkable">Try Canvas!</h2>
-				To try the EasyJS Canvas API, just add it to your page with just a few lines of code:
-				<code class="example block"><span class="tag">&ltscript</span> <span class="attribute">type</span>=<span class="value">"module"</span><span class="tag">></span>
-<span class="keyword">import</span> <span class="variable">Canvas</span> <span class="keyword">from</span> <span class="string">"http://gavinmorrow.com/easyjs/1/canvas/"</span>;
-<span class="comment">// You can set the version number (above) to any valid major version.</span>
-
-<span class="comment">// Your code here!</span>
-<span class="tag">&lt/script></span></code>
-				<h2 id="line-graph" class="linkable">Line Graph</h2>
-				The Line Graph API helps you make a line graph on the HTML canvas. To import it, you can import the rest of the Canvas interface, or just add this line: 
-				<code class="example">
-					<span class="keyword">import</span> <span class="variable">LineGraph</span> <span class="keyword">from</span> <span class="string">"http://gavinmorrow.com/easyjs/1/canvas/lineGraph/</span>;
-				</code>
-				<br/>
-				<br/>
-				To create a graph, 
-			</section>
-			<section id="math" class="section" data-text-transform="capitalize">
-				<h1 id="math-title" class="section-title linkable">Math</h1>
-			</section>
-			<section id="string" class="section" data-text-transform="capitalize">
-				<h1 id="string-title" class="section-title linkable">String</h1>
-			</section>
-			<section id="ui" class="section" data-text-transform="uppercase">
-				<h1 id="ui-title" class="section-title linkable">UI</h1>
-			</section>
-		</div>
-	</article>
-	
-	<script src="script.js" type="module"></script>
-</body>
-</html>
